BatchTrasactionProject/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-apache-cassandra
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-apache-spark==4.11.3

numpy
pandas
kaggle 
cassandra-driver 
openpyxl 
pyspark 
psycopg2-binary